Duplicate Domino Detection Issue Solved
The issue of dominos being detected more than once has been fixed. I
only have two pictures to test on, however it is working well for those,
when it wasnt before. This code also includes the previous update which
lowered the sensitivity to close domino placement. Multiple dominos can
only trigger as one domino now if the corners are very close (before it
was very loose), and they are alligned. (Bryden)